http: implement lockingcache to reduce lodestone queries
LockingCache is a middleware which caches responses from other middleware in-memory.
Additionally, it behaves in such a way that if a requests has not yet been resolved, and a subsequent request that would hit in cache arrives, the second request waits for the first to resolve and is then served from cache.
This behavior allows clients to request the same data multiple times in parallel, but hides such parallelism from the lodestone.